Improve App structure
