+++ src/App.js
@@ -1,25 +1,14 @@
 import {
-  NavLink, Route, Routes,
+  Route, Routes,
 } from 'react-router-dom';
 import Calculator from './components/Calculator';
 import Home from './components/Home';
+import NavBar from './components/NavBar';
 import Quote from './components/Quote';
 
 const MathMagiciansApp = () => (
   <>
-    <header className="top-section">
-      <h1>Math Magicians</h1>
-      <nav>
-        <NavLink
-          to="/"
-          className="link"
-        >
-          Home
-        </NavLink>
-        <NavLink className="link" to="/calculator">Calculator</NavLink>
-        <NavLink className="link" to="/quote">Quote</NavLink>
-      </nav>
-    </header>
+    <NavBar />
     <Routes>
       <Route path="/" element={<Home />} />
       <Route path="calculator" element={<Calculator />} />
